Send dict instead of User to the receivers.
Also, sender for user_deleted signal is None (instead of User object).
When running on Python 3, when blinker iterates through all receivers listening for sender (receivers_for() method on Signal), it tries to get hashable_identity for the sender.
To do that, it checks if sender has '__func__' attribute. This calls _ensure_data() on User object that we have just deleted (so it doesn't exist any more).